+++ src/sharedApi/kotlin/net/multigesture/kanama/api/OfflineMultiplayerPeer.kt
@@ -18,6 +18,7 @@ class OfflineMultiplayerPeer(handle: GodotHandle) : MultiplayerPeer(handle) {
         internal fun wrap(handle: RawSegment): OfflineMultiplayerPeer? =
             if (handle.address() == 0L) null else OfflineMultiplayerPeer(GodotHandle(handle))
 
+        // Instantiate an OfflineMultiplayerPeer.
         @JvmStatic
         fun create(): OfflineMultiplayerPeer =
             OfflineMultiplayerPeer(GodotHandle(ObjectCalls.constructObject("OfflineMultiplayerPeer")))