+++ src/sharedApi/kotlin/net/multigesture/kanama/api/Sprite2D.kt
@@ -370,6 +370,7 @@ class Sprite2D(handle: GodotHandle) : Node2D(handle) {
         internal fun wrap(handle: RawSegment): Sprite2D? =
             if (handle.address() == 0L) null else Sprite2D(GodotHandle(handle))
 
+        // Instantiate a Sprite2D.
         @JvmStatic
         fun create(): Sprite2D =
             Sprite2D(GodotHandle(ObjectCalls.constructObject("Sprite2D")))